requirements_dev.txt
mdutils==1.3.1
simple-term-menu==1.4.1
pyperclip==1.8.2
googlesearch-python==1.1.0
lxml>=4.8.0
html5lib>=1.1
beautifulsoup4>=4.9.0
pandas>=1.4.1
bump2version==0.5.11
flake8==3.7.8
tox==3.14.0
pytest==6.2.4
black==21.7b0
isort==5.10.1
bandit==1.7.2
pre-commit==2.17.0
